pyproject.toml
[project]
name = "typegpt"
version = "0.1.0"
description = "TypeGPT is a Python application that allows you to invoke various AI's and LLM's from any text field in your operating system."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "google-generativeai>=0.8.4",
    "pillow>=11.1.0",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "pyperclip>=1.9.0",
    "requests>=2.32.3",
]
